rename admin/expanding-formset.js (using kebab-case file)
