+++ client/webpack.config.js
@@ -33,7 +33,7 @@ module.exports = function exports() {
       'core',
       'date-time-chooser',
       'draftail',
-      'expanding_formset',
+      'expanding-formset',
       'filtered-select',
       'hallo-bootstrap',
       'hallo-plugins/hallo-hr',
