+++ wagtail/admin/templates/wagtailadmin/pages/_editor_js.html
@@ -20,7 +20,7 @@
 <script src="{% versioned_static 'wagtailadmin/js/comments.js' %}"></script>
 <script src="{% versioned_static 'wagtailadmin/js/vendor/rangy-core.js' %}"></script>
 <script src="{% versioned_static 'wagtailadmin/js/vendor/mousetrap.min.js' %}"></script>
-<script src="{% versioned_static 'wagtailadmin/js/expanding_formset.js' %}"></script>
+<script src="{% versioned_static 'wagtailadmin/js/expanding-formset.js' %}"></script>
 <script src="{% versioned_static 'wagtailadmin/js/modal-workflow.js' %}"></script>
 <script src="{% versioned_static 'wagtailadmin/js/page-editor.js' %}"></script>
 <script src="{% versioned_static 'wagtailadmin/js/vendor/xregexp.min.js' %}"></script>
