+++ wagtail/users/templates/wagtailusers/groups/includes/group_form_js.html
@@ -1,5 +1,5 @@
 {% load wagtailadmin_tags %}
-<script src="{% versioned_static 'wagtailadmin/js/expanding_formset.js' %}"></script>
+<script src="{% versioned_static 'wagtailadmin/js/expanding-formset.js' %}"></script>
 <script src="{% versioned_static 'wagtailadmin/js/modal-workflow.js' %}"></script>
 <script src="{% versioned_static 'wagtailusers/js/group-form.js' %}"></script>
